pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "german_verbs"
version = "0.1.0"
description = "German irregular verbs data and utilities"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.scripts]
german-verbs = "german_verbs.cli:main"
yaml2md = "german_verbs.converter:yaml_to_md_cli"

[tool.hatch.build.targets.wheel]
packages = ["german_verbs"] 